Added fix for 8-channel Gateways
Added AT+FSB=1 to GenericATcommands[]

Per Dave Kjendall's email:
AT+FSB=1 will set the device to the first 8 channels, which will match the conduit.
-Dave

Dave-
Thanks for the fast response. How can we check whether the mDots are configured for 8 vs 64 channels? We’re using the default software for the mDots, so I expect that they’re configured for 64 channels.  
Best
Mike


Are you still trying to use the 64 channel plan?  The conduit only supports 8 channels and the some of the radio modules have issues supporting the commands to restrict the channel plan (which still would take a dozen or so attempts by the device to be attempted. 
Dave
Senet, Inc